Add code to support the weird scaling quirks in OBS
When monitor DPI is high, window sources in OBS don't work right.

They capture the window size as the "real" pixel size on the monitor, but the window contents as the pixel size reported by windows.

This code makes the required adjustments to support this. at Scale=1, the default, all scaling methods have no effect.