Config Files
Basic start to being able to position controls based on a config file.  Idea is to have one person create the 2/4 runner settings in OBS, create a config file based on that and then export the scene info from OBS and zip all resources.  Once extracted on the restreamers end, the scene itself should be all setup with the correct positioning/bounding box info.